move all resources related settings to its own file
